requirements.txt
streamlit==1.41.1
pymupdf==1.25.1
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
numpy==2.1.3
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
openai==1.58.1
tiktoken==0.8.0
